refactor(dto): reorganize project structure into modules relocate DTOs class and mappers
