+++ src/main/java/mx/gob/cimientosdelrenacimiento/CimientosDelRenacimientoBackend/user/dto/UserRequestDTO.java
@@ -1,25 +1,22 @@
-package mx.gob.cimientosdelrenacimiento.CimientosDelRenacimientoBackend.dto;
+package mx.gob.cimientosdelrenacimiento.CimientosDelRenacimientoBackend.user.dto;
 
 import lombok.Data;
 
 @Data
-public class UserDTO {
+public class UserRequestDTO {
 
-    private Long idUser;
     private String name;
     private String middleName;
     private String firstLastName;
     private String secondLastName;
     private String phone;
     private String email;
+    private String password;
     private Boolean active;
     private Boolean isFirstLogin;
     private Boolean twoFactorEnabled;    
     private String twoFactorSecret;    
     private Integer verificationCode;
-    private RoleDTO role;
-    private Boolean deleted;
-    private String createdAt;
-    private String updatedAt;
-    private String deletedAT;
+    private Long roleId;
+
 }
